scripts/requirements.txt
google-cloud-vision==3.7.0
Pillow==10.2.0
PyYAML==6.0.1
requests==2.31.0
python-dateutil==2.8.2
pillow-heif==0.15.0
